add MADCTL options for display rotation
This adds a python script to compute parameters for the ST7789V 36h
MADCTL command to control screen rotation. The new comments in the
st789v devicetree file explain how to use MADCTL (aka "mdac") in
combination with x-offset, y-offset, width, and height properties
to get 0, 90, 180, or 270 degree screen rotations.