add subscribe after connect
This adds a function to subscribe to the MQTT topic once a flag is
set by the CONNACK event handler. I'm trying to structure this so I
have the option of moving time-consuming network operations to a
background thread or work queue. But, for now, I'm just calling
them from the event loop. I suppose it doesn't really matter if I
let the LVGL GUI stall since I'm not doing much with animations.