implement PUBLISH event handler
This responds to incoming published messages by checking that the
topic matches and then parsing the payload data. Expected payload
values are "1" or "0", which are meant to be compatible with
AdafruitIO dashboard switch widgets.

This does not include any GUI stuff for indicating the topic's
status on the TFT display. The PUBLISH event handler sets flags
for 1 or 0. The main event loop will need to respond to those
flags, but I haven't implemented that part yet.