improve startup status indicators
Instead of starting off with a disabled toggle switch at boot, this
now starts with a big "Waiting for MQTT Connect (check shell)"
message in the center of the screen. Also, instead of a gray
warning icon in the status bar, it now starts with a gray wifi
icon.

When wifi connects, the wifi icon in the status bar goes green.
When MQTT connects, the big "Waiting ..." message gets replaced
with the toggle switch. If the connections go down, the toggle
switch visibility and status indicators adjust accordingly.