publish toggle state for button press
This makes it so that pressing the Boot button will change the
switch widget's value and also PUBLISH the new value over MQTT.

I ended up moving the keypad press event to the LVGL active screen
instead of the switch itself. I initially tried to use the switch's
CHECKED property to track my toggle switch state. But, that didn't
work well at all. It seemed like I was getting a race condition
between MQTT (which echos published messages) and LVGL (which
animates switch toggling). Now, the main event loop is the only
thing that changes the switch state, and it considers both MQTT
PUBLISH messages and button presses when it does that (avoiding
race conditions).