read settings into context struct at boot
This loads all the wifi and MQTT settings stuff into the ZCtx
context struct. The `aio connect` shell command will use the MQTT
settings (taking the place of `aio conf ...`). *But*, the wifi ssid
and psk context struct fields are not used used for anything yet. I
need to figure out how to set up wifi with the network manager API
instead of relying on the `wifi connect ...` shell command.