refactor: untangle MQTT data structures
Previously, some of the important data structures for the MQTT
connection were spread around different files. The code for
manipulating the MQTT broker hostname, username, password, and
topic was also a bit scattered. This consolidates all that stuff
into zq3_mqtt.h and zq3_mqtt.c.

This is part of getting ready to implement TLS.