clean up TLS memory tuning stuff
This cleans up some of my leftover debugging stuff from tuning
memory allocation to get TLS working. RAM is a bit tight with both
LVGL and TLS, but the current arrangement seems to work fine.

As far as I can tell, the Zephyr support for moving wifi network
buffers to external SPIRAM does not currently work for ESP32-S3.
SPIRAM brokenness seems to be a known issue. There are currently a
handful of open SPIRAM-related issues in the zephyr repo.